ethtool: Add collector to E2E tests
